requirements.txt
absl-py==1.2.0
anyio==3.6.1
asttokens==2.0.5
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==7.1.2
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
easydict==1.9
et-xmlfile==1.1.0
executing==0.9.1
fastapi==0.85.0
filterpy==1.4.5
Flask==1.1.4
fonttools==4.34.4
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
ipython==8.4.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
kiwisolver==1.4.4
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
natsort==8.2.0
norfair==2.0.0
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
Pygments==2.12.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sniffio==1.3.0
stack-data==0.3.0
starlette==0.20.4
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.8.1
